+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Version 1.1.0
 Trickster Rework
+* Mythic Abilities
+	* Greater Enduring Spells
+		* Now extends spells that last 10+ minutes instead of 5+ minutes.
 
 ## Version 1.0.0
 This has been split and now depends on [TabletopTweaks-Core](https://github.com/Vek17/TabletopTweaks-Core/releases).